\docType{data}
\name{exampleDataset}
\alias{exampleDataset}
\title{Example Dataset}
\description{
This line could include a brief description of the data
}
\details{
The variables included in the dataset are: \itemize{
\item\code{Variable1} A vector of random numbers
\item\code{Variable2} Another vector of random numbers }
}

