\docType{class}
\name{polyreg-class}
\alias{initialize,}
\alias{polyreg-class}
\alias{polyreg-method}
\title{A Poly Regression object}
\description{
Object of class \code{polyreg} are created by the
\code{reg} functions
}
\details{
An object of the class `polyreg' has the following slots:
\itemize{ \item{X}{The covariates matrix input}
\item{Y}{The dependent variable numeric vector input}
\item{coefficient}{A matrix of coefficients in each model}
\item{R.squared}{A vector of the value of R-sqaured for
each model} }
}
\author{
Myunghoon Kang: \email{myunghoon@wustl.edu}
}

