\docType{package}
\name{PolyregPack}
\alias{PolyregPack}
\alias{PolyregPack-package}
\title{PolyregPack}
\description{
The PolyregPack package performs regression for all
possible combinations of covariates.
}
\examples{
\dontrun{
demo(reg)
demo(signi)
}
}
\author{
Myunghoon Kang: \email{myunghoon@wustl.edu}
}

